+++ backend/canister_installer/src/main.rs
@@ -23,6 +23,7 @@ async fn main() {
         market_maker: opts.market_maker,
         neuron_controller: opts.neuron_controller,
         escrow: opts.escrow,
+        translations: opts.translations,
         nns_root: opts.nns_root,
         nns_governance: opts.nns_governance,
         nns_internet_identity: opts.nns_internet_identity,
@@ -92,6 +93,9 @@ struct Opts {
 
     #[arg(long)]
     escrow: CanisterId,
+
+    #[arg(long)]
+    translations: CanisterId,
 
     #[arg(long)]
     nns_root: CanisterId,