+++ backend/canisters/identity/impl/src/guards.rs
@@ -1,9 +1,9 @@
 use crate::read_state;
 
-pub fn caller_is_governance_principal() -> Result<(), String> {
-    if read_state(|state| state.is_caller_governance_principal()) {
+pub fn caller_is_user_index_canister() -> Result<(), String> {
+    if read_state(|state| state.is_caller_user_index_canister()) {
         Ok(())
     } else {
-        Err("Caller is not the governance principal".to_owned())
+        Err("Caller is not the user_index canister".to_owned())
     }
 }