+++ backend/external_canisters/sns_root/c2c_client/src/lib.rs
@@ -1,0 +1,5 @@
+use canister_client::generate_candid_c2c_call;
+use sns_root_canister::*;
+
+// Queries
+generate_candid_c2c_call!(get_sns_canisters_summary);